+++ _builderApp/app/meta.php
@@ -1,6 +1,6 @@
 <?php
 $meta = array(
 	'appName' => 'BuilderApp',
-	'appVersion' => '0.5.0'
+	'appVersion' => '0.6.0'
 );
 ?>