+++ _builderApp/settings.php
@@ -1,5 +1,6 @@
 <?php
 
+// Minification
 $minify = true;
 
 // Single Variables
@@ -7,6 +8,23 @@ $singleVariables = array(
 	// Sample Variables
 	'siteName' => 'BuilderApp',
 	'authorUrl' => 'http://buzzingpixel.com'
+);
+
+// Remove Orphaned Files
+$removeOrphaned = true;
+
+// Exclude single files from orphan deletion
+$orphanedFileExlude = array(
+	// Sample Orphaned Excludes
+	'myfile.txt',
+	'mydir/myfile.txt'
+);
+
+// Exlude directories from orphaned deletion
+$orphanedDirectoryExclude = array(
+	// Sample Orphaned Directory Excludes
+	'mydir',
+	'myotherdir/mysubdir'
 );
 
 ?>